add option to gen seperate class for relations
Prismas standard query result is a class without relations. When using
the generated classes for response types in Nestjs it is very
unpractical to omit all relations in a dto manually or to have to pick
non relational fields. I created the option this way so it won't break
the api. If you leave out the option the generator works just like it
used to. If you set the seperateRelationFields option to true, it
generates two classes for each model, one without relations, and one
only containing them. This way it is easier to use mixins.