Fix inconsistency between `export` and `import`
The class template and related code export the class as simply `x.name`.
However, related entities are forced to `pascalCase` in import, which
breaks the import.

This probably works fine if you're doing the Right Thing™ by prisma and
naming your models in pascalCase, but it should break in all other
cases.